pyproject.toml
[project]
name = "banana-pencil-club"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.33.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
    "streamlit-javascript>=0.1.5",
]
